feat: implementação de estrutura para List seguindo modelo MVC

+++ back-php/src/repositories/ListRepository.php
@@ -2,8 +2,9 @@
 
 require_once(__DIR__ . '/../database/Connection.php');
 
-class ListRepository{
-    
+class ListRepository
+{
+
     // POST https://lunodoro/lista
     public static function insertIntoListFromDatabase($id_usuario, $nome_lista, $descricao, $id_tipo_lista)
     {
@@ -64,10 +65,24 @@ class ListRepository{
     public static function updateList($id_usuario, $id, $nome_lista, $descricao, $id_tipo_lista)
     {
         try {
+            $existingList = ListRepository::findListFromDatabase($id_usuario, $id);
+
+            if (!$existingList) {
+                throw new Exception("Lista não encontrada", 404);
+            }
+
+            if (
+                $existingList['nome_lista'] === $nome_lista &&
+                $existingList['descricao'] === $descricao &&
+                $existingList['id_tipo_lista'] == $id_tipo_lista
+            ) {
+                throw new Exception("Nenhuma mudança detectada nos dados da lista.", 400);
+            }
+
             $conn = Connection::getConnection();
             $stmt = $conn->prepare("UPDATE lista 
-                                    SET nome_lista = ?, descricao = ?, id_tipo_lista = ? 
-                                    WHERE id = ? AND id_usuario = ?");
+                                SET nome_lista = ?, descricao = ?, id_tipo_lista = ? 
+                                WHERE id = ? AND id_usuario = ?");
             $stmt->execute([
                 $nome_lista,
                 $descricao,
@@ -75,16 +90,24 @@ class ListRepository{
                 $id,
                 $id_usuario
             ]);
+
             return $stmt->rowCount();
         } catch (PDOException $e) {
             throw new Exception("Erro ao atualizar a lista", 500);
         }
     }
 
+
     // DELETE https://lunodoro/usuarios/{id_usuario}/lista/{id}
     public static function removeList($id_usuario, $id)
     {
         try {
+            $existingList = ListRepository::findListFromDatabase($id_usuario, $id);
+
+            if (!$existingList) {
+                throw new Exception("Lista não encontrada", 404);
+            }
+
             $conn = Connection::getConnection();
             $stmt = $conn->prepare("DELETE FROM lista WHERE id_usuario = ? AND id = ?");
             $stmt->execute([
@@ -97,4 +120,3 @@ class ListRepository{
         }
     }
 }
-?>
